Refactor Goal Management Components
- Removed the existing GoalSection component and replaced it with a new structure that utilizes GoalCard for better modularity and reusability.
- Introduced GoalCard component to encapsulate goal display logic, including header, content, and actions.
- Created GoalSection component to manage the overall display of goals, statistics, and loading/error states.
- Added types for goals and their associated data structures in a new types.ts file for better type safety and clarity.
- Updated the index file to export the new components and types for easier imports.
