+++ frontend/src/features/chat/components/bubbles/bot/goals/index.ts
@@ -1,0 +1,4 @@
+export { default as GoalSection } from "./GoalSection";
+export { GoalCard } from "./GoalCard";
+export * from "./types";
+export { default } from "./GoalSection";
